requirements.txt
transformers
langchain
bitsandbytes
accelerate
langchain_experimental==0.0.54
langchain_community
docx2txt
pypdf
llama-cpp-python
faiss-cpu